refactor/full-frontend: +auth controller
