+++ gambitto-frontend/src/services/index.ts
@@ -1,0 +1,4 @@
+export * from "./ChessService";
+export * from "./FriendshipService";
+export * from "./UserService";
+export * from "./AuthService";
