+++ gambitto-frontend/src/store/index.ts
@@ -1,0 +1,2 @@
+export * from "./reducers";
+export * from "./store";
