+++ gambitto-frontend/src/store/reducers/index.ts
@@ -1,0 +1,3 @@
+export * from "./notificationsSlice";
+export * from "./userSlice";
+export * from "./wsSlice";
